Changes to give only unique duplicates for divisions.
